+++ lib/syntax.ml
@@ -1,7 +1,7 @@
-type t = Normal | Cram | Mli
+type t = Markdown | Cram | Mli
 
 let pp fs = function
-  | Normal -> Fmt.string fs "normal"
+  | Markdown -> Fmt.string fs "markdown"
   | Cram -> Fmt.string fs "cram"
   | Mli -> Fmt.string fs "mli"
 
@@ -10,12 +10,12 @@ let equal x y = x = y
 let infer ~file =
   match Filename.extension file with
   | ".t" -> Some Cram
-  | ".md" -> Some Normal
+  | ".md" -> Some Markdown
   | ".mli" -> Some Mli
   | _ -> None
 
 let of_string = function
-  | "markdown" | "normal" -> Some Normal
+  | "markdown" | "normal" -> Some Markdown
   | "cram" -> Some Cram
   | "mli" -> Some Mli
   | _ -> None